+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Add new with_timeout constructor allowing to define connection/runtime timeouts
 - Upgrade snafu to 0.6
 - Use backtraces in errors, can be enabled using `ts3-query = { features = "backtrace" }`
+- Accept &str/String/&String in raw escape helpers
+- *Breaking*: `whoami` takes a paramter for unescaping all values
 
 ### 0.1.4
 - Add `raw::parse_multi_hashmap` to handle `clientlist` like commands.